+++ admin/sql/uninstall.mysql.utf8.sql
@@ -21,6 +21,7 @@ DROP TABLE IF EXISTS `#__componentbuilder_fieldtype`;
 DROP TABLE IF EXISTS `#__componentbuilder_language_translation`;
 DROP TABLE IF EXISTS `#__componentbuilder_language`;
 DROP TABLE IF EXISTS `#__componentbuilder_server`;
+DROP TABLE IF EXISTS `#__componentbuilder_repository`;
 DROP TABLE IF EXISTS `#__componentbuilder_help_document`;
 DROP TABLE IF EXISTS `#__componentbuilder_admin_fields`;
 DROP TABLE IF EXISTS `#__componentbuilder_admin_fields_conditions`;