+++ admin/src/Model/ComponentbuilderModel.php
@@ -67,7 +67,7 @@ class ComponentbuilderModel extends ListModel
 		$icons  = [];
 		// view groups array
 		$viewGroups = array(
-			'main' => array('png.compiler', 'png.joomla_components', 'png.joomla_modules', 'png.joomla_plugins', 'png.powers', 'png.search', 'png.admin_views', 'png.custom_admin_views', 'png.site_views', 'png.template.add', 'png.templates', 'png.layouts', 'png.dynamic_get.add', 'png.dynamic_gets', 'png.custom_codes', 'png.placeholders', 'png.libraries', 'png.snippets', 'png.validation_rules', 'png.field.add', 'png.fields', 'png.fields.catid_qpo0O0oqp_com_componentbuilder_po0O0oq_field', 'png.fieldtypes', 'png.fieldtypes.catid_qpo0O0oqp_com_componentbuilder_po0O0oq_fieldtype', 'png.language_translations', 'png.languages', 'png.servers', 'png.help_documents')
+			'main' => array('png.compiler', 'png.joomla_components', 'png.joomla_modules', 'png.joomla_plugins', 'png.powers', 'png.search', 'png.admin_views', 'png.custom_admin_views', 'png.site_views', 'png.template.add', 'png.templates', 'png.layouts', 'png.dynamic_get.add', 'png.dynamic_gets', 'png.custom_codes', 'png.placeholders', 'png.libraries', 'png.snippets', 'png.validation_rules', 'png.field.add', 'png.fields', 'png.fields.catid_qpo0O0oqp_com_componentbuilder_po0O0oq_field', 'png.fieldtypes', 'png.fieldtypes.catid_qpo0O0oqp_com_componentbuilder_po0O0oq_fieldtype', 'png.language_translations', 'png.languages', 'png.servers', 'png.repositories', 'png.help_documents')
 		);
 		// view access array
 		$viewAccess = [
@@ -184,6 +184,11 @@ class ComponentbuilderModel extends ListModel
 			'server.access' => 'server.access',
 			'servers.submenu' => 'server.submenu',
 			'servers.dashboard_list' => 'server.dashboard_list',
+			'repository.create' => 'repository.create',
+			'repositories.access' => 'repository.access',
+			'repository.access' => 'repository.access',
+			'repositories.submenu' => 'repository.submenu',
+			'repositories.dashboard_list' => 'repository.dashboard_list',
 			'help_document.create' => 'help_document.create',
 			'help_documents.access' => 'help_document.access',
 			'help_document.access' => 'help_document.access',