+++ libraries/vendor_jcb/VDM.Joomla/src/Abstraction/Grep.php
@@ -189,16 +189,22 @@ abstract class Grep implements GrepInterface
 		{
 			foreach ($this->paths as $n => &$path)
 			{
-				if (isset($path->owner) && strlen($path->owner) > 1 &&
-						isset($path->repo) && strlen($path->repo) > 1)
+				if (isset($path->organisation) && strlen($path->organisation) > 1 &&
+						isset($path->repository) && strlen($path->repository) > 1)
 				{
 					// build the path
-					$path->path = trim($path->owner) . '/' . trim($path->repo);
+					$path->path = trim($path->organisation) . '/' . trim($path->repository);
 
 					// update the branch
-					if ($path->branch === 'default' || empty($path->branch))
+					if ($path->read_branch === 'default' || empty($path->read_branch))
 					{
-						$path->branch = null;
+						$path->read_branch = null;
+					}
+
+					// only update the write branch if set
+					if (isset($path->write_branch) && ($path->write_branch === 'default' || empty($path->write_branch)))
+					{
+						$path->write_branch = null;
 					}
 
 					// set local path